requirements.txt
numpy
tensorflow-gpu
keras
h5py
coremltools
